add BLM support
