+++ fflogs_rotation/job_data/potencies/magical_ranged/BlackMage/7_0-BlackMage.csv
@@ -1,22 +1,21 @@
 job,ability_id,ability_name,level,buff_id,base_potency,damage_type,combo_bonus,combo_potency,positional_bonus,positional_potency,combo_positional_bonus,combo_positional_potency
-BlackMage,141,Fire,90,,180,direct,,,,,,
-BlackMage,152,Fire III,90,,260,direct,,,,,,
-BlackMage,3577,Fire IV,90,,310,direct,,,,,,
-BlackMage,25794,High Fire II,90,,140,direct,,,,,,
-BlackMage,16507,Xenoglossy,90,,880,direct,,,,,,
-BlackMage,16505,Despair,90,,340,direct,,,,,,
-BlackMage,25797,Paradox,90,,500,direct,,,,,,
-BlackMage,153,Thunder III,90,,50,direct,,,,,,
-BlackMage,153,Thunder III,90,1000164,400,direct,,,,,,
-BlackMage,1000163,Thunder III (tick),90,,35,magic-dot,,,,,,
-BlackMage,7420,Thunder IV,90,,50,direct,,,,,,
-BlackMage,7420,Thunder IV,90,1000164,170,direct,,,,,,
-BlackMage,1001210,Thunder IV (tick),90,,20,magic-dot,,,,,,
-BlackMage,142,Blizzard,90,,180,direct,,,,,,
-BlackMage,154,Blizzard III,90,,260,direct,,,,,,
-BlackMage,3576,Blizzard IV,90,,310,direct,,,,,,
-BlackMage,25795,High Blizzard II,90,,140,direct,,,,,,
-BlackMage,159,Freeze,90,,120,direct,,,,,,
-BlackMage,156,Scathe,90,,100,direct,,,,,,
-BlackMage,162,Flare,90,,220,direct,,,,,,
-BlackMage,7422,Foul,90,,600,direct,,,,,,
+BlackMage,141,Fire,100,,180,direct,,,,,,
+BlackMage,152,Fire III,100,,280,direct,,,,,,
+BlackMage,3577,Fire IV,100,,310,direct,,,,,,
+BlackMage,25794,High Fire II,100,,100,direct,,,,,,
+BlackMage,16507,Xenoglossy,100,,880,direct,,,,,,
+BlackMage,16505,Despair,100,,340,direct,,,,,,
+BlackMage,25797,Paradox,100,,500,direct,,,,,,
+BlackMage,142,Blizzard,100,,180,direct,,,,,,
+BlackMage,154,Blizzard III,100,,280,direct,,,,,,
+BlackMage,3576,Blizzard IV,100,,310,direct,,,,,,
+BlackMage,25795,High Blizzard II,100,,100,direct,,,,,,
+BlackMage,159,Freeze,100,,120,direct,,,,,,
+BlackMage,156,Scathe,100,,100,direct,,,,,,
+BlackMage,162,Flare,100,,240,direct,,,,,,
+BlackMage,7422,Foul,100,,600,direct,,,,,,
+BlackMage,36989,Flare Star,100,,400,direct,,,,,,
+BlackMage,36986,High Thunder,100,,200,direct,,,,,,
+BlackMage,1003871,High Thunder (tick),100,,55,magic-dot,,,,,,
+BlackMage,36987,High Thunder II,100,,100,direct,,,,,,
+BlackMage,1003872,High Thunder II (tick),100,,40,magic-dot,,,,,,
